chapter 1/requirements.txt
strands-agents
strands-agents-tools
langchain
langchain-aws
langgraph
